+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/main/java/mozilla/components/browser/engine/gecko/GeckoEngine.kt
@@ -333,6 +333,10 @@ class GeckoEngine(
                     runtime.settings.fontSizeFactor = it
                 }
             }
+
+        override var forceUserScalableContent: Boolean
+            get() = runtime.settings.forceUserScalableEnabled
+            set(value) { runtime.settings.forceUserScalableEnabled = value }
     }.apply {
         defaultSettings?.let {
             this.javascriptEnabled = it.javascriptEnabled
@@ -349,6 +353,7 @@ class GeckoEngine(
             this.suspendMediaWhenInactive = it.suspendMediaWhenInactive
             this.fontInflationEnabled = it.fontInflationEnabled
             this.fontSizeFactor = it.fontSizeFactor
+            this.forceUserScalableContent = it.forceUserScalableContent
         }
     }
 }